add get all and get by id order

+++ src/main/java/com/enigma/orderin/service/impl/OrderDetailServiceImpl.java
@@ -29,7 +29,7 @@ public class OrderDetailServiceImpl implements OrderDetailService {
         }
 
         return OrderDetailResponse.builder()
-                .orderDetailId(lastInsertId.intValue())
+                .orderDetailId(lastInsertId)
                 .product(ProductResponse.builder()
                         .productId(orderDetail.getProductDetail().getProduct().getId())
                         .price(orderDetail.getProductDetail().getPrice())
